--- src/main/java/jp/co/topgate/sekiguchi/kai/MessageProcesser.java
@@ -1,11 +1,0 @@
-package jp.co.topgate.sekiguchi.kai;
-
-/**
- * Messageに関するコントローラー
- * Created by sekiguchikai on 2016/11/22.
- */
-public class MessageProcesser {
-    public byte[] handle(HTTPRequest httpRequest, HTTPResponse httpResponse) {
-        return null;
-    }
-}
