+++ src/main/java/jp/co/topgate/sekiguchi/kai/MessageProcessor.java
@@ -1,0 +1,13 @@
+package jp.co.topgate.sekiguchi.kai;
+
+import java.io.File;
+
+/**
+ * Messageに関するコントローラー
+ * Created by sekiguchikai on 2016/11/22.
+ */
+class MessageProcessor implements Processor {
+    public void process(HTTPRequest httpRequest, HTTPResponse httpResponse) {
+
+    }
+}
