--- src/main/java/jp/co/topgate/sekiguchi/kai/Processer.java
@@ -1,8 +1,0 @@
-package jp.co.topgate.sekiguchi.kai;
-
-/**
- * Created by sekiguchikai on 2016/11/22.
- */
-public interface Processer {
-    void handle();
-}
