+++ src/main/java/jp/co/topgate/sekiguchi/kai/Processor.java
@@ -1,0 +1,9 @@
+package jp.co.topgate.sekiguchi.kai;
+
+/**
+ *
+ * Created by sekiguchikai on 2016/11/22.
+ */
+ interface Processor {
+    void process(HTTPRequest httpRequest, HTTPResponse httpResponse);
+}
